Add Spot contract tests
Add documented Spot fixtures, public live captures, contract tests, request construction tests, and public integration smoke tests.

Update Spot response enum parsing, public insurance history handling, and Unified currency discount-tier parsing for current public responses.

+++ src/Gate.IO.Api/Spot/Responses/GateSpotInsurance.cs
@@ -15,7 +15,7 @@ public record GateSpotInsurance
     /// Balance
     /// </summary>
     [JsonProperty("balance")]
-    public long Balance { get; set; }
+    public decimal Balance { get; set; }
 
     /// <summary>
     /// Creation time, timestamp, milliseconds
